Finalise documentation and test for regex patterns
We're now able to use regular expression patterns in addition to strings
when searching a piece of text. This means you can either search for
`fox` or `/(fox)/` with the two being equivalent.

But you can also have more complex queries, with more control over the
regular expression used. See the tests for more examples.

fixes #17